refactor: TokenListProvider only load the balance
